refactor test code
refactor test code

1. api-tests is a combo program, contains three parts: test cases, test server(robot) and an easy to use test launcher.
2. Test cases and test server can be run separately use different command line options.
3. Test launcher just an easy way to run test server and test cases in local mode.
4. Use socket instead of stdin/stdout as the control channel between test cases and test server. So the test cases and the test server can be run locally or remotely, and this change will make the api-tests more flexible with different network environment.
5. The test launcher should still use stdout/stderr gather the outputs and the logs, aggregate them into curses windows or write to files.
6. Scenario 1: run tests on different machine. e.g.
   Test server command: api-tests --mode=server --listen=0.0.0.0:8070
   Test cases command: api-tests --mode=client --server=192.168.8.10:8070
7. Scenario 2: run tests on single machine use easy way
   Command: api-tests

+++ Carrier/apps/shell/shell.c
@@ -31,7 +31,6 @@
 #include <limits.h>
 #include <inttypes.h>
 #include <sys/stat.h>
-#include <pthread.h>
 
 #ifdef HAVE_UNISTD_H
 #include <unistd.h>
@@ -69,7 +68,7 @@
 #endif
 
 #include <rc_mem.h>
-
+#include <pthread.h>
 #include <ela_carrier.h>
 #include <ela_session.h>
 
@@ -948,7 +947,7 @@ static void stream_bulk_receive(ElaCarrier *w, int argc, char *argv[])
         int duration = (int)((end.tv_sec - start.tv_sec) * 1000000 +
                              (end.tv_usec - start.tv_usec)) / 1000;
         duration = (duration == 0)  ? 1 : duration;
-        float speed = ((session_ctx.bytes / duration) * 1000) / 1024;
+        float speed = (float)((session_ctx.bytes / duration) * 1000) / 1024;
 
         output("\nFinish! Total %" PRIu64 " bytes in %d.%03d seconds. %.2f KB/s\n",
                session_ctx.bytes,
@@ -1306,7 +1305,7 @@ static void *bulk_write_thread(void *arg)
     duration = (int)((end.tv_sec - start.tv_sec) * 1000000 +
                      (end.tv_usec - start.tv_usec)) / 1000;
     duration = (duration == 0) ? 1 : duration;
-    speed = (((args->packet_size * args->packet_count) / duration) * 1000) / 1024;
+    speed = (float)(((args->packet_size * args->packet_count) / duration) * 1000) / 1024;
 
     output("\nFinish! Total %" PRIu64 " bytes in %d.%03d seconds. %.2f KB/s\n",
            (uint64_t)(args->packet_size * args->packet_count),
